DB-Access-Agent/requirements.txt
google-generativeai
SQLAlchemy